requirements.txt
fastmcp>=0.1.0
shodan>=1.28.0
asyncio
